+++ Sources/Classes/Headers/Public/RSEventRepository.h
@@ -54,7 +54,7 @@ NS_ASSUME_NONNULL_BEGIN
     RSEventFilteringPlugin *eventFilteringPlugin;
     NSLock* lock;
     dispatch_source_t source;
-    dispatch_queue_t queue;
+    dispatch_queue_t repositoryQueue;
     RSClient *client;
 }